# server-openapi.yaml
openapi: 3.1.0
info:
  title: FIRMUPS Server API
  version: '0.1.0'
  description: |
    The firmups server API allows for device and update management.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
...
servers:
  - url: "http://127.0.0.1:3000"
    description: Local development server
...
paths:
...
  /firmware/{id}:
    get:
      tags:
        - Firmware
      security:
        - api_key: []
      summary: Get firmware
      operationId: getFirmware
      parameters:
        - name: id
          in: path
          description: ID of the Firmware to be returned
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Firmware'
        '404':
          description: Firmware not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalError"
    delete:
      tags:
        - Firmware
      security:
        - api_key: []
      summary: Delete firmware
      operationId: deleteFirmware
      parameters:
        - name: id
          in: path
          description: ID of the Firmware to be deleted
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Firmware'
        '404':
          description: Firmware not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalError"
...
components:
  schemas:
...
    Firmware:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        version:
          type: string
        file_id:
          type: string
        size:
          type: string
        sha256:
          type: string
      required:
        - name
        - version
        - file_id
        - size
        - sha256
...
    InternalError:
      description: Masked internal error. The id can be matched with the backend logs.
      type: object
      properties:
        error_id:
          type: string
      required:
        - error_id
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
  securitySchemes:
    api_key:
      type: apiKey
      name: x-api-key
      in: header